requirements.txt
python-dotenv~=0.18.0
slack_bolt
requests
PyMuPDF